feat(global_controls): add audio slider mock
